apod_service/requirements.txt
Django>=4.2
djangorestframework>=3.14
requests>=2.28
python-dotenv>=1.0
cachetools>=5.3
django-cors-headers>=4.0
# Optional (for production scaled cache):
# django-redis>=5.2
